ocpb-124 refactor add init code for signUpAdmin
